requirements.txt
streamlit==1.34.0
langchain==0.0.312
tiktoken==0.7.0
openai==0.28.1
python-dotenv==1.0.0
slack_bolt==1.20.0
slack_sdk==3.31.0
